tests/wpt/tests/client-hints/viewport-size/viewport-size-iframe.https.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(t => {
  return new Promise(resolve => {
    window.addEventListener("message", t.step_func(e => {
      assert_equals(e.data.viewportWidth, window.innerWidth.toString());
      assert_equals(e.data.viewportHeight, window.innerHeight.toString());
      resolve();
    }));
  });
});
</script>
<iframe src="../resources/viewport-frame.py" width=503 height=614></iframe>
